chore: move @compas/insight to @compas/stdlib and @compas/server
Closes #850
