--- packages/insight/.npmignore
@@ -1,2 +1,0 @@
-node_modules/
-*.test.js
