--- packages/insight/index.js
@@ -1,13 +1,0 @@
-export { newLogger } from "./src/logger/logger.js";
-
-export { bytesToHumanReadable, printProcessMemoryUsage } from "./src/memory.js";
-
-export {
-  newEvent,
-  eventStart,
-  eventRename,
-  eventStop,
-  newEventFromEvent,
-} from "./src/events.js";
-
-export { postgresTableSizes } from "./src/postgres.js";
