--- packages/insight/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@compas/insight",
-  "version": "0.0.132",
-  "description": "Simple logger in NDJSON format",
-  "main": "./index.js",
-  "exports": "./index.js",
-  "types": "./index.d.ts",
-  "type": "module",
-  "keywords": [
-    "compas",
-    "logger"
-  ],
-  "license": "MIT",
-  "maintainers": [
-    {
-      "name": "Dirk de Visser",
-      "email": "dirkdev98@gmail.com"
-    },
-    {
-      "name": "Daniël Hansen"
-    }
-  ],
-  "homepage": "https://compasjs.com",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/compasjs/compas.git",
-    "directory": "packages/insight"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "engines": {
-    "node": ">=15"
-  }
-}
