+++ packages/stdlib/index.js
@@ -37,3 +37,15 @@ export {
   filenameForModule,
   dirnameForModule,
 } from "./src/utils.js";
+
+export { newLogger } from "./src/logger/logger.js";
+
+export { bytesToHumanReadable, printProcessMemoryUsage } from "./src/memory.js";
+
+export {
+  newEvent,
+  eventStart,
+  eventRename,
+  eventStop,
+  newEventFromEvent,
+} from "./src/events.js";
