+++ packages/stdlib/src/events.js
