+++ packages/stdlib/src/logger/logger.bench.js
