+++ packages/stdlib/src/logger/writer.js
