+++ packages/stdlib/src/memory.test.js
