+++ packages/store/package.json
@@ -14,7 +14,6 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "@compas/insight": "0.0.132",
     "@compas/stdlib": "0.0.132",
     "@types/minio": "7.0.7",
     "mime-types": "2.1.30",
